scanner: remove unnecessary cast
'tok' is already an int now.

+++ src/xkbcomp/scanner.c
@@ -169,7 +169,7 @@ skip_more_whitespace_and_comments:
 
         /* Keyword. */
         tok = keyword_to_token(s->buf);
-        if ((int) tok != -1) return tok;
+        if (tok != -1) return tok;
 
         yylval->str = strdup(s->buf);
         if (!yylval->str)
